local variable names, factories moved
