+++ Assets/Scripts/OrchestraArmy/Entity/Entities/Enemies/Regular/Factories.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f872c5a1dbedb774fb3b4e8856243f27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
