+++ Assets/Scripts/OrchestraArmy/Entity/Entities/Enemies/Regular/Factories/FluteEnemyFactory.cs
@@ -1,4 +1,4 @@
-namespace OrchestraArmy.Entity.Entities.Enemies.Regular
+namespace OrchestraArmy.Entity.Entities.Enemies.Regular.Factories
 {
     public class FluteEnemyFactory : IEnemyFactory
     {
