+++ Assets/Scripts/OrchestraArmy/Entity/Entities/Enemies/Regular/Factories/GuitarEnemyFactory.cs
@@ -1,4 +1,4 @@
-namespace OrchestraArmy.Entity.Entities.Enemies.Regular
+namespace OrchestraArmy.Entity.Entities.Enemies.Regular.Factories
 {
     public class GuitarEnemyFactory : IEnemyFactory
     {
