+++ Assets/Scripts/OrchestraArmy/Entity/Entities/Enemies/Regular/Factories/SousaphoneEnemyFactory.cs
@@ -1,4 +1,4 @@
-namespace OrchestraArmy.Entity.Entities.Enemies.Regular
+namespace OrchestraArmy.Entity.Entities.Enemies.Regular.Factories
 {
     public class SousaphoneEnemyFactory : IEnemyFactory
     {
